LLM/RagIntegrate/requirements.txt
psycopg2
psycopg2-binary
sentence-transformers
pgvector
pyyaml
faiss-cpu
langchain
ollama
